bioinformatics-mcp-search/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.15
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
google-generativeai>=0.3.1
chromadb>=0.4.22
langchain-community>=0.0.13
langchain-core>=0.1.10
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
requests>=2.31.0
aiohttp>=3.9.1
biopython>=1.83
bioconductor>=0.1.0
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1 